+++ src/main/java/com/swisscom/tasks/task3/repository/RoleRepository.java
@@ -1,6 +1,6 @@
 package com.swisscom.tasks.task3.repository;
 
-import com.swisscom.tasks.task3.model.Role;
+import com.swisscom.tasks.task3.model.auth.Role;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
